feat: add `defaultGroupByItems` to cost-analysis-page-store
